update scripts.js
alert for clear storage button

+++ app/static/scripts/scripts.js
@@ -267,5 +267,9 @@ function gameDefeat() {
 
 function clearStorage() {
     // Clear localStorage items 
-    localStorage.clear();
+    if (confirm("WARNING: confirm statistics reset")) {
+        localStorage.clear();
+      } else {
+        //
+      }
   }
